KAFKA-3594; After calling MemoryRecords.close() method, hasRoomFor() method should return false
This exception is occurring when producer is trying to append a record to a Re-enqueued record batch in the accumulator. We should not allow to add a record to Re-enqueued record batch. This is due a bug in MemoryRecords.java/hasRoomFor() method. After calling MemoryRecords.close() method, hasRoomFor() method should return false.

Author: Manikumar reddy O <manikumar.reddy@gmail.com>

Reviewers: Ismael Juma, Grant Henke, Guozhang Wang

Closes #1249 from omkreddy/KAFKA-3594